Refactors the verify methods to their own files
